Use parseFloat instead of parseInt

+++ modules/gtp/command.js
@@ -2,7 +2,7 @@ var helper = require('../helper')
 
 var Command = function(id, name, arguments) {
     this.internalId = helper.getId()
-    this.id = parseInt(id)
+    this.id = parseFloat(id)
     this.name = name
     this.arguments = arguments ? arguments : []
 }
